requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==1.10
django-debug-toolbar==1.9.1
djangorestframework==3.7.7
idna==2.6
lxml==4.1.1
pkg-resources==0.0.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
pytest==3.4.2
pytest-django==3.1.2
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
